requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
scikit-learn
requests
python-multipart
pydantic
pillow
torch
torchvision